LayoutTests/platform/efl/fast/regions/content-webkit-from-flow-parsing-expected.txt

Tests parsing of -webkit-flow-from property

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testCSSText("-webkit-flow-from: first-flow") is "first-flow"
PASS testCSSText("-webkit-flow-from: none") is "none"
PASS testCSSText("-webkit-flow-from: ") is ""
PASS testCSSText("-webkit-flow-from: 'first-flow'") is ""
PASS testCSSText("-webkit-flow-from: 1") is ""
PASS testCSSText("-webkit-flow-from: 1.2") is ""
PASS testCSSText("-webkit-flow-from: -1") is ""
PASS testCSSText("-webkit-flow-from: 12px") is ""
PASS testCSSText("-webkit-from-flow: first flow") is ""
PASS testComputedStyle("first-flow") is "first-flow"
PASS testComputedStyle("'first flow'") is "none"
PASS testComputedStyle("none") is "none"
PASS testComputedStyle("initial") is "none"
PASS testComputedStyle("inherit") is "none"
PASS testComputedStyle("1") is "none"
PASS testComputedStyle("1.2") is "none"
PASS testComputedStyle("-1") is "none"
PASS testComputedStyle("12px") is "none"
PASS successfullyParsed is true

TEST COMPLETE
